mk/checkconf.mk: allow spaces and double quotes in CFG_ config variables
Fixes issues with the check-conf-h and check-conf-mk functions, which
would error out or generate incorrect output on strings containing spaces
and/or double quotes.

The single quote character is used in the shell commands that builds
the list of values, instead of the double quote, so that there is no
conflict with double quotes appearing in the variables themselves. As a
consequence, single quotes cannot appear anywhere in the output string.
So remove the comments that contained quotes, and which were not that
useful anyway.

Test case:

 $ grep CFG_TEST core/arch/arm/plat-vexpress/conf.mk
 CFG_TEST1 ?= "test string"
 CFG_TEST2 ?= test string
 CFG_TEST3 ?= "test"
 CFG_TEST4 ?= test
 $ make -s
 $ grep CFG_TEST out/arm-plat-vexpress/conf.mk
 CFG_TEST1="test string"
 CFG_TEST2=test string
 CFG_TEST3="test"
 CFG_TEST4=test
 $ grep CFG_TEST out/arm-plat-vexpress/include/generated/conf.h
 #define CFG_TEST1 "test string"
 #define CFG_TEST2 test string
 #define CFG_TEST3 "test"
 #define CFG_TEST4 test

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Tested-by: Zeng Tao <prime.zeng@hisilicon.com>