mm: drop the forced PL1 RW permissions for user TAs
We have used simple memory access permission model in OP-TEE, if PL1
permission is forced RW, the PL0 permission can only be set to RW or
no access, so the PL0 permission is set to RW in the user TAs which is
not as expected.

Fix it as follow,
1. when TA is in loading process, the PL1 is set to own the RW
permission while the PL0 with no accesss.
2. when the TA is loaded, the PL0 is set to own the required
permissions defined in ta elf program headers and from the access
control model, the PL1 will own the same permissions as the PL0.

Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Zeng Tao <prime.zeng@hisilicon.com>