core: bugfix tee_mmu_rem_rwmem()
Prior to this patch tee_mmu_rem_rwmem() checked that the size of the
supplied mobj matches the size of the region entry. Since a mobj may
have any size and region entries always are page or page directory
sized, region entries that should have been removed were not. This patch
fixes that by only checking that mobj and va matches the region entry.

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>