core: do not assume GNU awk is available
The rshift(), strtonum() and and() functions are GNU extensions to awk.
Therefore, the build will fail if another variant of awk is installed
on the build system.
Replace the awk calls with bash arithmetic to avoid any issues. We know
we can rely on the shell being bash, since the main Makefile has
SHELL = /bin/bash.

Fixes: aa4cc147a462 ("core: fix build error with large tee-init_load_addr")
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>