mobj: add mobj_reg_shm and mobj_mapped_shm
mobj_reg_shm represents registered shared memory. This is basically
a list of pages provided by normal world. It can be used to pass
memory parameters to TAs.

mobj_mapped_shm is built on top of mobj_reg_shm. It is almost
the same thing, but it is mapped to OP-TEE virtual address space,
so OP-TEE kernel can access such buffers.

Signed-off-by: Volodymyr Babchuk <vlad.babchuk@gmail.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>