core: FS: introduce dirfile interface
Introduces the dirfile interface which keeps a list of secure storage
objects. Each object is represented by the UUID of the owning TA, a hash
of the object, and a handle to the name used in normal world to store
the encrypted data of the object.

The interface allows queued atomic updates to avoid difficult races
during creation and renaming of objects.

By keeping the list of the secure object files in a single database the
collected state of the objects can be represented by a single hash or
even a counter. This gives some flexibility when implementing
anti-rollback protection.

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>