core: pager: avoid page faults with pager lock held
When a normal thread takes the pager lock there must be no page faults
while the lock is held. This is has so far been ensured by feeding the
function to the KEEP_PAGER() macro, but this will only take care of the
actual code not stack usage. This patch introduces a new pager lock
function, pager_lock_check_stack() which will make sure that the
estimated needed stack is mapped before the lock is acquired.

Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>