docs: add status icons indicating supported platforms
This adds a set of new icons (svg-files) used to tell the current status
of platforms that are officially supported in OP-TEE (found in
README.md). Green means that they are up-to-date and has been tested
on the latest OP-TEE release. Orange means that it was not tested on the
latest release and red are the ones who has not been tested for quite a
while (for the red ones we are also adding the last known release they
where tested against).

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>