libutee: fix TA_CreateEntryPoint() and TA_DestroyEntryPoint()
Fixes issues observed with keep alive single instance TAs:
- TA_CreateEntryPoint() must only be called once,
- TA_DestroyEntryPoint() must not be called when the last session is
closed, because the instance is still alive.

While we're at it, simplify the code a bit by not using a separate
variable (ta_ref_count) to track whether we have sessions or not.
Simply use TAILQ_EMPTY() on the session queue instead.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Kevin Peng <kevinp@marvell.com>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>