plat-rcar_gen4: auth: Fix key memory deficiency when ICUM FW decrypt, compare the content_cert
With the size of current memory < TA_VERIFICATION_SIZE=100000U>.
When making a request of service  "secureboot_dec_and_comp",
the ICUM FW returns an error code that is: SERV_RAM_SIZE_ERROR
after service completed.
This error code means the configuration of secure ram too small
for the number of keys to manage.
Therefore, we have to increase the size of contained key memory
to avoid this issue.