plat-rcar_gen4: Add LSI support
There are three LSI options of board which are supported on gen4 currently. These are S4, V4M and V4H
We have to build with an option LSI=xxx, 'xxx' means:
 [1] LSI = V4H < for V4H target board >
 [2] LSI = V4M < for V4M target board >
 [3] LSI = S4 < for S4 target board >
If user don't provide an LSI value for using, the system will be built with LSI = S4 as default option automatically.

Signed-off-by: Nhi Nguyen <nhi.nguyen.kx@renesas.com>
Signed-off-by: Dien Pham <dien.pham.ry@renesas.com>
Signed-off-by: Tu Tran <tu.tran.wz@renesas.com>