common: Initialize to support for LPM (Low Power Mode)
This patch initializes to support Low Power Mode.
It helps to perform LPM booting successfully.

In LPM, power of CA cores are powered off but the
memory is kept. Next booting, CA core will be
triggered by CR core to re-boot.

TODO: in this patch, disabling of checking is used,
but it is better to reset and re-initialize (global/
static variable). This is next improvement item.

Signed-off-by: Dien Pham <dien.pham.ry@renesas.com>